Fix/ecoms-7825 adjusts gti trade in checkout when delete products with trade in

+++ checkout-ui-custom/src/components/_tradeIn.js
@@ -15,8 +15,9 @@ export default class TradeIn {
         : ''
 
     const transport = getTransport ? JSON.parse(getTransport) : ''
-    if (items.length && transport.length && localStorage.getItem('transport')) {
-      await this.checkTradeIn(items, transport)
+
+    if (items.length && transport.length) {
+      this.checkTradeIn(items, transport)
     } else if (
       !items.length &&
       transport.length &&
@@ -32,7 +33,7 @@ export default class TradeIn {
     return window.__RUNTIME__.rootPath ? window.__RUNTIME__.rootPath : ''
   }
 
-  async checkTradeIn(items, transport) {
+  checkTradeIn(items, transport) {
     let totalTradeIn = 0
 
     if (transport.length) {
@@ -71,7 +72,7 @@ export default class TradeIn {
     } else if (totalTradeIn === 0) {
       $('#total-details-tradein').remove()
       $('#text-details-tradein').remove()
-      await this.removeCustomDataTradeIn()
+      this.removeCustomDataTradeIn()
 
       return
     }
@@ -140,17 +141,21 @@ export default class TradeIn {
 
   async removeCustomDataTradeIn() {
     const { orderFormId } = window.vtexjs.checkout.orderForm
-    localStorage.removeItem('transport')
-
-    await $.ajax({
-      url: `${this.rootPath()}/v1/pub/deleteCheckoutCustomData/${orderFormId}/domain/trade_in_option_selected`,
-      type: 'DELETE',
-    })
-
-    await $.ajax({
-      url: `${this.rootPath()}/v1/pub/deleteCheckoutCustomData/${orderFormId}/domain/trade_in_total_value`,
-      type: 'DELETE',
-    })
+    const openTextField = localStorage.getItem('tradeInCustom')
+    if(openTextField !== null || openTextField !== "null"){
+      localStorage.removeItem('tradeInCustom')
+      localStorage.removeItem('transport')
+      window.vtexjs.checkout.sendAttachment('openTextField', { value: null })
+      await $.ajax({
+        url: `${this.rootPath()}/v1/pub/deleteCheckoutCustomData/${orderFormId}/domain/trade_in_option_selected`,
+        type: 'DELETE',
+      })
+  
+      await $.ajax({
+        url: `${this.rootPath()}/v1/pub/deleteCheckoutCustomData/${orderFormId}/domain/trade_in_total_value`,
+        type: 'DELETE',
+      })
+    }
   }
 
   async validateTradeinCustomData() {
